Create marloweC4nftLoanContract
This Marlowe smart contract facilitates a Borrower to deposit an NFT for a Loan of 10,000 $C4 at a 4.5% interest rate for a 3-day term (for simplicity). 

If the Borrower does not pay back the interest to the Lender daily, (e.g. they miss a single interest payment before the daily timeout) the contract will be terminated and the Lender will get to keep the NFT.